Beam: JobHandler: a simple job-oriented rpc layer over beam streams.
Docker-DCO-1.1-Signed-off-by: Solomon Hykes <solomon@docker.com> (github: shykes)
